+++ app/src/main/java/com/waminiyi/realestatemanager/core/model/data/Estate.kt
@@ -4,7 +4,7 @@ data class Estate(
     val uuid: String, // Unique identifier for the property
     val type: EstateType, // Type of the property (e.g., Apartment, House, etc.)
     val price: Int, // Price of the property in dollars
-    val area: Float,
+    val area: Int,
     val mainPhoto: Photo, // Main property image
     val addressCity: String, // Address of the property
 )
